Mybatis에 대한 전반적인 로직 변경 (조회부터 삭제까지 사용자가 결과를 볼수 있도록)

+++ mybatis/src/main/java/io/simpolor/mybatis/domain/Student.java
@@ -4,17 +4,17 @@ import java.util.List;
 
 public class Student {
 
-	private int seq;
+	private long seq;
 	private String name;
 	private int grade;
 	private int age;
 	private List<String> hobby;
 
-	public int getSeq() {
+	public long getSeq() {
 		return seq;
 	}
 
-	public void setSeq(int seq) {
+	public void setSeq(long seq) {
 		this.seq = seq;
 	}
